split up animation files
