+++ client/components/aframe/animations/commands/index.js
@@ -1,0 +1,3 @@
+import speak from './speak'
+
+export default { speak }
