+++ client/components/aframe/animations/emotions/index.js
@@ -1,0 +1,3 @@
+import joy from './joy'
+
+export default { joy }
